reafctor : rename constant file name and update import statements of constants.
