+++ src/app/robots.ts
@@ -1,4 +1,4 @@
-import { APP_URL } from "@/constants/Constants";
+import { APP_URL } from "@/constants";
 import type { MetadataRoute } from "next";
 
 export default function robots(): MetadataRoute.Robots {
