+++ src/app/sitemap.ts
@@ -1,13 +1,13 @@
-import { APP_URL } from "@/constants/Constants";
+import { APP_URL } from "@/constants";
 import type { MetadataRoute } from "next";
 
 export default function sitemap(): MetadataRoute.Sitemap {
   return [
     {
       url: APP_URL,
-      lastModified: new Date().toISOString(), 
-      changeFrequency: "weekly", 
-      priority: 1.0, 
+      lastModified: new Date().toISOString(),
+      changeFrequency: "weekly",
+      priority: 1.0,
     },
   ];
 }
