+++ src/constants/index.ts
@@ -1,0 +1,1 @@
+export { APP_DETAILS, APP_URL, METADATA } from "./app.constants";
